+++ tests/lib/audio/decode-audio.worker.browser.ts
@@ -12,8 +12,8 @@ test("decodes audio file and returns StoredAudioData", async () => {
 
   expect(result.length).toBe(217728);
   expect(result.numberOfChannels).toBe(2);
-  expect(result.channels[0]).toBeInstanceOf(Float32Array);
-  expect(result.channels[1]).toBeInstanceOf(Float32Array);
+  expect(result.channels[0]).toBeInstanceOf(Int16Array);
+  expect(result.channels[1]).toBeInstanceOf(Int16Array);
   expect(result.sampleRate).toBe(48000);
 });
